requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
sentence_transformers
langchain_huggingface
pypdf
bs4
arxiv
pymupdf
wikipedia
lxml
langchain-openai
faiss-cpu
streamlit
langchain_groq
duckduckgo-search
validators
unstructured
youtube-transcript-api
pytube
numexpr
langchain_nvidia_ai_endpoints